Math: change return type of 'math_sin' and 'math_cos' to i32
All functions of the math module now return i32 values.

+++ include/math.h
@@ -66,8 +66,8 @@ INLINE i32 math_brad(i32 angle) {
 }
 
 // output in range [-0x4000, +0x4000]
-extern i16 math_sin(u16 angle);
+extern i32 math_sin(u16 angle);
 
-INLINE i16 math_cos(u16 angle) {
+INLINE i32 math_cos(u16 angle) {
     return math_sin(angle + math_brad(90));
 }
